Rename py23 to compatibility
This is a more common name for this file in standard projects.
