+++ sparse/compatibility.py
